Create typeSpeed.py
{source text for the sample.txt file that feeds the typing speed test input.}
This typing test is based on two python packages constructing the contents and serving them up in a random order.
The lines you see here are written in a specific order. However the source code deploys Python's inbuilt split(/n) method in order to split each line of this paragraph.
For every line, we create a quasi-array object element that is held as a singular element in the paragraph.
Then, the source code uses an imported Math.random package.
The Math.random.choice() method undertakes a randomised sorting of the inpuitted total data which is this paragraph treated as if it were an array.
Then each element of that array, which is the individual sentences demarcated by the /n newline separator become jumbled up and served up to you so you can test your typing speed without writing the sentences in the same order.